+++ releasenotes.md
@@ -46,6 +46,7 @@
 * Controllers now rumble slightly when the player picks something up. This feature can be toggled using the new `joy_rumble_pickup` CVAR, which is `on` by default and `off` when vanilla mode is enabled.
 * A bug is fixed whereby the `vid_fullscreen` CVAR couldn’t be changed in the console.
 * The player’s weapon is now displayed correctly if wide enough to touch the right side of the screen and the `vid_capfps` CVAR is not `35`.
+* The `weapon` CVAR now allows values `1` to `7`.
 
 ![](https://github.com/bradharding/www.doomretro.com/raw/master/wiki/bigdivider.png)